requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
dj-database-url==2.1.0
mysqlclient==2.2.4
requests==2.31.0
